Promote ov006/dMgJump3DMario_c and convert 15 members to methods
Fold the 28 per-symbol shards covering ov006 0x020c762c..0x020c8a30 into
src/actors/dMgJump3DMario_c.cpp -- the translation unit the linker run says
they were -- and, in the same edit, turn fifteen of them into real
dMgJump3DMario_c:: methods.

28/28 MATCH. objisolate clean, relocation destinations clean, and all 28
functions emit in the expected ROM-ascending section order under
`#pragma defer_codegen off`. Twenty of the twenty-eight members are now real
methods: the three virtuals, D1 and C1 already carried mangled names, and
fifteen more converted here.

The other eight stay free `func_ov006_*` functions, and the linker enumerated
that ceiling rather than a guess: twelve direct `bl` sites in seven
still-unpromoted `.c` shards immediately below the run name them by their C
spelling. Two independent scans -- relocs.txt filtered on `module:overlay(6)`,
and a raw ARM BL decode of extracted/overlays/overlay_0006.bin -- agreed on the
same 12 sites and the same 8 targets. They convert when those shards are
promoted, not before.

symbols.txt moves in this commit, which a text-only promotion normally never
does. Fifteen 8-byte pointer-to-member records sit in ov006's unowned .data at
0x0213b020..0x0213b098, and dsd resolves each record's code word by the NAME
symbols.txt gives at the target address. A converted member whose row still
said `func_ov006_*` would link its .data word as 0x00000000 with a green
compile, and no byte gate at `verify` catches it. Renaming the member is one
edit, not two.

The fifteen new member names are COINED. The cartridge proves each member's
address, size, that its first argument is the object, and -- for the nine
reached through a pointer-to-member -- which record installs it and which
member writes that record. It proves no spelling. The derivation is written out
in include/dMgJump3DMario_c.h and the fifteen rows are logged in
symbols/actor_renames.tsv, each separating what the ROM proves from what was
chosen. The three virtuals stay address-derived; nothing in the ROM carries
their 2004 spelling.

include/dMgJump3DMario_c.h gains fifteen non-virtual declarations and no field
and no virtual. Byte-neutrality is proven, not asserted: the two already-promoted
minigame TUs that include it, the base constructor TU and all five of this
class's own shards were compiled with the pinned mwccarm 2004/b56 before and
after; all eight object sha256s are identical.

compiler_only_output is 10 rows: six vague-linkage RTTI/vtable records for this
class and its abstract base, every one homed in ov006's own symbols.txt and so
dispositioned deadstrip-data with a canonical address, and four homeless text
symbols (D2, D0, C2 and the base D2) dispositioned deadstrip. externalized_output
is empty.

Link evidence. `tubuild.py linkcheck ov006/dMgJump3DMario_c` refuses at [4/8]
with "intact production requires one .text claim and at least one non-text
claim" -- the expected, benign refusal for a text-only promotion. The real
control is `linkcheck --baseline --module ov006` run against this promoted tree,
which is the only self-consistent state given that the rename and the promotion
are atomic: module ov006 built 525344 bytes against retail 525344, module
fidelity 106/106 exact at 100.000000%, ROM-build analysis PASS, and the ROM came
out at sha256 d1506e90efae5e2d2cf119926a4ac2a291bd5ca78349d09d5024e1a918c478e8
-- the pristine baseline hash. Every renamed .data record word therefore
resolved to a real address. The 9 `dsd check symbols --fail` errors in that run
are all in ARM9 main and ITCM, none names ov006, and a --baseline run performs
no TU substitution, so they belong to the tree.

One measured byte diff and its fix: func_ov006_020c81e0's shard spelled the
8-byte record it copies as `struct { int a, b; }`. C block-moved it; C++
scalarises two named ints and schedules the load after the store, two words
differing at +0x64 and +0x68. Spelling it with an array member restored 28/28.

CONVERTED ratchet banked with --update: two paths left the set
(src/_ZN16dMgJump3DMario_cC1Ev.cpp, src/_ZN16dMgJump3DMario_cD1Ev.cpp) and five
member rows entered it, all five in this class. Nothing unrelated moved.

The queue row's already_promoted column is re-derived from queue_audit's own
output for this row only; four unrelated rows that the generator also wants to
rewrite are pre-existing drift and are left to the integrator.
