+++ config/arm9/overlays/ov006/delinks.txt
@@ -622,117 +622,9 @@ src/func_ov006_020c7574.c:
     complete
     .text start:0x020c7574 end:0x020c762c
 
-src/_ZN16dMgJump3DMario_c12Unk_020c762cEv.cpp:
+src/actors/dMgJump3DMario_c.cpp:
     complete
-    .text start:0x020c762c end:0x020c76d0
-
-src/_ZN16dMgJump3DMario_c12Unk_020c76d0Ev.cpp:
-    complete
-    .text start:0x020c76d0 end:0x020c76d8
-
-src/_ZN16dMgJump3DMario_c12Unk_020c76d8Ev.cpp:
-    complete
-    .text start:0x020c76d8 end:0x020c76e0
-
-src/func_ov006_020c76e0.cpp:
-    complete
-    .text start:0x020c76e0 end:0x020c7734
-
-src/func_ov006_020c7734.cpp:
-    complete
-    .text start:0x020c7734 end:0x020c7860
-
-src/func_ov006_020c7860.cpp:
-    complete
-    .text start:0x020c7860 end:0x020c78ec
-
-src/func_ov006_020c78ec.cpp:
-    complete
-    .text start:0x020c78ec end:0x020c79a8
-
-src/func_ov006_020c79a8.cpp:
-    complete
-    .text start:0x020c79a8 end:0x020c7a30
-
-src/func_ov006_020c7a30.c:
-    complete
-    .text start:0x020c7a30 end:0x020c7ba4
-
-src/func_ov006_020c7ba4.c:
-    complete
-    .text start:0x020c7ba4 end:0x020c7c68
-
-src/func_ov006_020c7c68.c:
-    complete
-    .text start:0x020c7c68 end:0x020c802c
-
-src/func_ov006_020c802c.c:
-    complete
-    .text start:0x020c802c end:0x020c8048
-
-src/func_ov006_020c8048.c:
-    complete
-    .text start:0x020c8048 end:0x020c8084
-
-src/func_ov006_020c8084.cpp:
-    complete
-    .text start:0x020c8084 end:0x020c814c
-
-src/func_ov006_020c814c.cpp:
-    complete
-    .text start:0x020c814c end:0x020c81e0
-
-src/func_ov006_020c81e0.c:
-    complete
-    .text start:0x020c81e0 end:0x020c8270
-
-src/func_ov006_020c8270.c:
-    complete
-    .text start:0x020c8270 end:0x020c833c
-
-src/func_ov006_020c833c.c:
-    complete
-    .text start:0x020c833c end:0x020c85a0
-
-src/func_ov006_020c85a0.c:
-    complete
-    .text start:0x020c85a0 end:0x020c85bc
-
-src/func_ov006_020c85bc.cpp:
-    complete
-    .text start:0x020c85bc end:0x020c862c
-
-src/func_ov006_020c862c.c:
-    complete
-    .text start:0x020c862c end:0x020c864c
-
-src/func_ov006_020c864c.c:
-    complete
-    .text start:0x020c864c end:0x020c8658
-
-src/func_ov006_020c8658.c:
-    complete
-    .text start:0x020c8658 end:0x020c8680
-
-src/func_ov006_020c8680.c:
-    complete
-    .text start:0x020c8680 end:0x020c8768
-
-src/func_ov006_020c8768.c:
-    complete
-    .text start:0x020c8768 end:0x020c87d0
-
-src/func_ov006_020c87d0.cpp:
-    complete
-    .text start:0x020c87d0 end:0x020c893c
-
-src/_ZN16dMgJump3DMario_cD1Ev.cpp:
-    complete
-    .text start:0x020c893c end:0x020c8a04
-
-src/_ZN16dMgJump3DMario_cC1Ev.cpp:
-    complete
-    .text start:0x020c8a04 end:0x020c8a30
+    .text start:0x020c762c end:0x020c8a30
 
 src/func_ov006_020c8a30.c:
     complete
