+++ config/arm9/overlays/ov006/symbols.txt
@@ -159,25 +159,25 @@ _ZN16dMgJump3DMario_c12Unk_020c76d8Ev kind:function(arm,size=0x8) addr:0x020c76d
 func_ov006_020c76e0 kind:function(arm,size=0x54) addr:0x020c76e0
 func_ov006_020c7734 kind:function(arm,size=0x12c) addr:0x020c7734
 func_ov006_020c7860 kind:function(arm,size=0x8c) addr:0x020c7860
-func_ov006_020c78ec kind:function(arm,size=0xbc) addr:0x020c78ec
-func_ov006_020c79a8 kind:function(arm,size=0x88) addr:0x020c79a8
-func_ov006_020c7a30 kind:function(arm,size=0x174) addr:0x020c7a30
-func_ov006_020c7ba4 kind:function(arm,size=0xc4) addr:0x020c7ba4
-func_ov006_020c7c68 kind:function(arm,size=0x3c4) addr:0x020c7c68
-func_ov006_020c802c kind:function(arm,size=0x1c) addr:0x020c802c
-func_ov006_020c8048 kind:function(arm,size=0x3c) addr:0x020c8048
+_ZN16dMgJump3DMario_c9StateDampEv kind:function(arm,size=0xbc) addr:0x020c78ec
+_ZN16dMgJump3DMario_c9EnterDampEv kind:function(arm,size=0x88) addr:0x020c79a8
+_ZN16dMgJump3DMario_c9StateHoldEv kind:function(arm,size=0x174) addr:0x020c7a30
+_ZN16dMgJump3DMario_c9EnterHoldEv kind:function(arm,size=0xc4) addr:0x020c7ba4
+_ZN16dMgJump3DMario_c9StateMoveEv kind:function(arm,size=0x3c4) addr:0x020c7c68
+_ZN16dMgJump3DMario_c9EnterMoveEv kind:function(arm,size=0x1c) addr:0x020c802c
+_ZN16dMgJump3DMario_c12StateFallOutEv kind:function(arm,size=0x3c) addr:0x020c8048
 func_ov006_020c8084 kind:function(arm,size=0xc8) addr:0x020c8084
-func_ov006_020c814c kind:function(arm,size=0x94) addr:0x020c814c
+_ZN16dMgJump3DMario_c12StateRiseOutEv kind:function(arm,size=0x94) addr:0x020c814c
 func_ov006_020c81e0 kind:function(arm,size=0x90) addr:0x020c81e0
-func_ov006_020c8270 kind:function(arm,size=0xcc) addr:0x020c8270
-func_ov006_020c833c kind:function(arm,size=0x264) addr:0x020c833c
-func_ov006_020c85a0 kind:function(arm,size=0x1c) addr:0x020c85a0
-func_ov006_020c85bc kind:function(arm,size=0x70) addr:0x020c85bc
+_ZN16dMgJump3DMario_c8EnterHitEv kind:function(arm,size=0xcc) addr:0x020c8270
+_ZN16dMgJump3DMario_c11StateBounceEv kind:function(arm,size=0x264) addr:0x020c833c
+_ZN16dMgJump3DMario_c11EnterBounceEv kind:function(arm,size=0x1c) addr:0x020c85a0
+_ZN16dMgJump3DMario_c11StateWindUpEv kind:function(arm,size=0x70) addr:0x020c85bc
 func_ov006_020c862c kind:function(arm,size=0x20) addr:0x020c862c
-func_ov006_020c864c kind:function(arm,size=0xc) addr:0x020c864c
+_ZN16dMgJump3DMario_c9StateIdleEv kind:function(arm,size=0xc) addr:0x020c864c
 func_ov006_020c8658 kind:function(arm,size=0x28) addr:0x020c8658
-func_ov006_020c8680 kind:function(arm,size=0xe8) addr:0x020c8680
-func_ov006_020c8768 kind:function(arm,size=0x68) addr:0x020c8768
+_ZN16dMgJump3DMario_c12StateRespawnEv kind:function(arm,size=0xe8) addr:0x020c8680
+_ZN16dMgJump3DMario_c12EnterRespawnEv kind:function(arm,size=0x68) addr:0x020c8768
 func_ov006_020c87d0 kind:function(arm,size=0x16c) addr:0x020c87d0
 _ZN16dMgJump3DMario_cD1Ev kind:function(arm,size=0xc8) addr:0x020c893c
 _ZN16dMgJump3DMario_cC1Ev kind:function(arm,size=0x2c) addr:0x020c8a04
