+++ config/converted-backslide-exceptions.jsonl
@@ -381,3 +381,5 @@
 {"path": "src/_ZN12dScMgPanel_cD1Ev.cpp", "reason": "ov006/dScMgPanel_c: 71 one-function shards absorbed into src/actors/dScMgPanel_c.cpp (whole-run TU promotion). The three banked CONVERTED shards -- _ZN12dScMgPanel_cD1Ev, _ZN12dScMgPanel_cD0Ev, _ZN12dScMgPanel_c6RenderEv -- resolve MOVED into the merged TU."}
 {"path": "src/_ZN15dScMgCurling2_cD0Ev.cpp", "reason": "ov006/dScMgCurling2_c segment-A TU promotion: the CONVERTED destructor shards src/_ZN15dScMgCurling2_cD1Ev.cpp and D0Ev.cpp are absorbed into src/actors/dScMgCurling2_c.cpp, whose other 29 members are still the byte-matched func_ov006_* recoveries. Readability is the humanizer's stage; byte match outranks it, and a reconstructed TU must spell vague-linkage symbols directly."}
 {"path": "src/_ZN15dScMgCurling2_cD1Ev.cpp", "reason": "ov006/dScMgCurling2_c segment-A TU promotion: the CONVERTED destructor shards src/_ZN15dScMgCurling2_cD1Ev.cpp and D0Ev.cpp are absorbed into src/actors/dScMgCurling2_c.cpp, whose other 29 members are still the byte-matched func_ov006_* recoveries. Readability is the humanizer's stage; byte match outranks it, and a reconstructed TU must spell vague-linkage symbols directly."}
+{"path": "src/_ZN16dMgJump3DMario_cC1Ev.cpp", "reason": "TU promotion ov006/dMgJump3DMario_c: the 28 per-symbol shards for this class were folded into src/actors/dMgJump3DMario_c.cpp, the translation unit the linker run says they were. Two of the 28 were in the CONVERTED baseline as whole-file identities -- src/_ZN16dMgJump3DMario_cC1Ev.cpp and src/_ZN16dMgJump3DMario_cD1Ev.cpp -- and both survive the move as passing MEMBER identities in the merged file, alongside three more members that pass for the first time (EnterMove, EnterBounce, StateIdle): 2 paths out, 5 member rows in, no readability lost. The BACKSLID rows read 'member regression' only because the destination file also carries members that do not pass, which is the arithmetic of a 28-member TU, not a regression: the eight members still named func_ov006_* are pinned by twelve direct bl sites in seven still-unpromoted .c shards below the run, the three virtuals are address-derived because the ROM carries no spelling for them, and the raw offset arithmetic is what the shards already had. Byte evidence unchanged and improved: 28/28 MATCH, objisolate clean, reloc destinations clean, and linkcheck --baseline on the promoted tree rebuilt the ROM at sha256 d1506e90efae5e2d2cf119926a4ac2a291bd5ca78349d09d5024e1a918c478e8."}
+{"path": "src/_ZN16dMgJump3DMario_cD1Ev.cpp", "reason": "TU promotion ov006/dMgJump3DMario_c: the 28 per-symbol shards for this class were folded into src/actors/dMgJump3DMario_c.cpp, the translation unit the linker run says they were. Two of the 28 were in the CONVERTED baseline as whole-file identities -- src/_ZN16dMgJump3DMario_cC1Ev.cpp and src/_ZN16dMgJump3DMario_cD1Ev.cpp -- and both survive the move as passing MEMBER identities in the merged file, alongside three more members that pass for the first time (EnterMove, EnterBounce, StateIdle): 2 paths out, 5 member rows in, no readability lost. The BACKSLID rows read 'member regression' only because the destination file also carries members that do not pass, which is the arithmetic of a 28-member TU, not a regression: the eight members still named func_ov006_* are pinned by twelve direct bl sites in seven still-unpromoted .c shards below the run, the three virtuals are address-derived because the ROM carries no spelling for them, and the raw offset arithmetic is what the shards already had. Byte evidence unchanged and improved: 28/28 MATCH, objisolate clean, reloc destinations clean, and linkcheck --baseline on the promoted tree rebuilt the ROM at sha256 d1506e90efae5e2d2cf119926a4ac2a291bd5ca78349d09d5024e1a918c478e8."}
