+++ config/match_provenance.jsonl
@@ -290,8 +290,8 @@
 {"id":"ov002:0x020d6dac","module":"ov002","addr":34434476,"name":"func_ov002_020d6dac","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov002_020d6dac.c"}
 {"id":"ov002:0x020e558c","module":"ov002","addr":34493836,"name":"_ZN6Player13InitResourcesEv","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/_ZN6Player13InitResourcesEv.cpp"}
 {"id":"ov006:0x020c94e0","module":"ov006","addr":34378976,"name":"func_ov006_020c94e0","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov006_020c94e0.cpp"}
-{"id":"ov006:0x020c7c68","module":"ov006","addr":34372712,"name":"func_ov006_020c7c68","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov006_020c7c68.c"}
-{"id":"ov006:0x020c7c68","module":"ov006","addr":34372712,"name":"func_ov006_020c7c68","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov006_020c7c68.c"}
+{"id":"ov006:0x020c7c68","module":"ov006","addr":34372712,"name":"func_ov006_020c7c68","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/actors/dMgJump3DMario_c.cpp"}
+{"id":"ov006:0x020c7c68","module":"ov006","addr":34372712,"name":"func_ov006_020c7c68","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/actors/dMgJump3DMario_c.cpp"}
 {"id":"ov006:0x020c94e0","module":"ov006","addr":34378976,"name":"func_ov006_020c94e0","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov006_020c94e0.cpp"}
 {"id":"ov006:0x02102274","module":"ov006","addr":34611828,"name":"func_ov006_02102274","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov006_02102274.c"}
 {"id":"ov074:0x02121800","module":"ov074","addr":34740224,"name":"func_ov074_02121800","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/func_ov074_02121800.c"}
@@ -599,17 +599,17 @@
 {"id":"ov075:0x021160b4","module":"ov075","addr":34693300,"name":"_ZN10dScEntry_c15graphCallback_cC1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN10dScEntry_c15graphCallback_cC1Ev.cpp"}
 {"id":"ov075:0x0211a838","module":"ov075","addr":34711608,"name":"_ZN10dScEntry_c6icon_cC1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN10dScEntry_c6icon_cC1Ev.cpp"}
 {"id":"ov006:0x020c4048","module":"ov006","addr":34357320,"name":"_ZN22dMg3DHeyhoObjAdapter_cC2Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN22dMg3DHeyhoObjAdapter_cC2Ev.cpp"}
-{"id":"ov006:0x020c893c","module":"ov006","addr":34375996,"name":"_ZN16dMgJump3DMario_cD1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN16dMgJump3DMario_cD1Ev.cpp"}
-{"id":"ov006:0x020c8a04","module":"ov006","addr":34376196,"name":"_ZN16dMgJump3DMario_cC1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN16dMgJump3DMario_cC1Ev.cpp"}
+{"id":"ov006:0x020c893c","module":"ov006","addr":34375996,"name":"_ZN16dMgJump3DMario_cD1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/actors/dMgJump3DMario_c.cpp"}
+{"id":"ov006:0x020c8a04","module":"ov006","addr":34376196,"name":"_ZN16dMgJump3DMario_cC1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/actors/dMgJump3DMario_c.cpp"}
 {"id":"ov006:0x020edec0","module":"ov006","addr":34528960,"name":"_ZN11dScMgJump_cD1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN11dScMgJump_cD1Ev.cpp"}
 {"id":"ov006:0x020edf54","module":"ov006","addr":34529108,"name":"_ZN11dScMgJump_cD0Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN11dScMgJump_cD0Ev.cpp"}
 {"id":"ov006:0x020eebe8","module":"ov006","addr":34532328,"name":"_ZN12dScMgJump2_cD1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN12dScMgJump2_cD1Ev.cpp"}
 {"id":"ov006:0x020eec9c","module":"ov006","addr":34532508,"name":"_ZN12dScMgJump2_cD0Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN12dScMgJump2_cD0Ev.cpp"}
 {"id":"ov006:0x020eeafc","module":"ov006","addr":34532092,"name":"MgBounceAndPounce_Spawn","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"ai-tdd-labs","srcPath":"src/MgBounceAndPounce_Spawn.cpp"}
 {"id":"ov006:0x020efaf0","module":"ov006","addr":34536176,"name":"dScMgJump2_c_Spawn","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/dScMgJump2_c_Spawn.c"}
-{"id":"ov006:0x020c762c","module":"ov006","addr":34371116,"name":"_ZN16dMgJump3DMario_c12Unk_020c762cEv","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"lunavyqo","srcPath":"src/_ZN16dMgJump3DMario_c12Unk_020c762cEv.cpp"}
-{"id":"ov006:0x020c76d0","module":"ov006","addr":34371280,"name":"_ZN16dMgJump3DMario_c12Unk_020c76d0Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN16dMgJump3DMario_c12Unk_020c76d0Ev.cpp"}
-{"id":"ov006:0x020c76d8","module":"ov006","addr":34371288,"name":"_ZN16dMgJump3DMario_c12Unk_020c76d8Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN16dMgJump3DMario_c12Unk_020c76d8Ev.cpp"}
+{"id":"ov006:0x020c762c","module":"ov006","addr":34371116,"name":"_ZN16dMgJump3DMario_c12Unk_020c762cEv","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"lunavyqo","srcPath":"src/actors/dMgJump3DMario_c.cpp"}
+{"id":"ov006:0x020c76d0","module":"ov006","addr":34371280,"name":"_ZN16dMgJump3DMario_c12Unk_020c76d0Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/actors/dMgJump3DMario_c.cpp"}
+{"id":"ov006:0x020c76d8","module":"ov006","addr":34371288,"name":"_ZN16dMgJump3DMario_c12Unk_020c76d8Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/actors/dMgJump3DMario_c.cpp"}
 {"id":"ov006:0x020e7a50","module":"ov006","addr":34503248,"name":"_ZN17dMg3DEspAnimSet_cD1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN17dMg3DEspAnimSet_cD1Ev.cpp"}
 {"id":"ov006:0x020e7aac","module":"ov006","addr":34503340,"name":"_ZN17dMg3DEspAnimSet_cC1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN17dMg3DEspAnimSet_cC1Ev.cpp"}
 {"id":"ov006:0x020e80d8","module":"ov006","addr":34504920,"name":"_ZN15dMg3DEspModel_cD1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN15dMg3DEspModel_cD1Ev.cpp"}
