--- src/_ZN16dMgJump3DMario_c12Unk_020c76d8Ev.cpp
@@ -1,8 +1,0 @@
-//cpp
-// @symbol _ZN16dMgJump3DMario_c12Unk_020c76d8Ev
-#include "dMgJump3DMario_c.h"
-
-void *dMgJump3DMario_c::Unk_020c76d8()
-{
-    return (char *)this + 0x14;
-}
