--- src/_ZN16dMgJump3DMario_cC1Ev.cpp
@@ -1,7 +1,0 @@
-//cpp
-// @symbol _ZN16dMgJump3DMario_cC1Ev
-#include "dMgJump3DMario_c.h"
-
-dMgJump3DMario_c::dMgJump3DMario_c()
-{
-}
