--- src/func_ov006_020c85a0.c
@@ -1,3 +1,0 @@
-struct S { int w[2]; };
-extern struct S data_ov006_0213b060;
-void func_ov006_020c85a0(char *p) { *(struct S *)(p + 0x3c) = data_ov006_0213b060; }
