--- src/func_ov006_020c862c.c
@@ -1,2 +1,0 @@
-struct S{int w[2];}; extern struct S data_ov006_0213b050;
-void func_ov006_020c862c(int* c, int v){ *(short*)((char*)c+0x32)=(short)v; *(struct S*)((char*)c+0x3c)=data_ov006_0213b050; }
