--- src/func_ov006_020c864c.c
@@ -1,4 +1,0 @@
-void func_ov006_020c864c(int *p)
-{
-    p[6] = 0;
-}
